+++ src/components/searchBar/SearchBar.module.css
@@ -5,12 +5,13 @@
     box-shadow: 10px 15px 20px rgba(0, 0, 0, 0.25);
     display: flex;
     flex: 1 1 auto;
-    margin: 25px 25px 25px 25px;
+    margin: 25px;
     box-sizing: border-box;
 }
 
 .input {
     border: none;
     width: 100%;
+    padding: 0 25px;
     border-radius: 10px;
 }
